move methods based on comments
